pyproject.toml
[tool.poetry]
name = "micro-backend"
version = "0.1.0"
description = "Django REST API micro-backend"
authors = ["Developer <dev@example.com>"]
readme = "README.md"
packages = [{include = "micro_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "5.1.6"
django-allauth = "64.1.0"
django-cors-headers = "^4.0.0"
djangorestframework = "^3.13.1"
djangorestframework-api-key = "^3.0.0"
djangorestframework-simplejwt = "^5.1.0"
dj-rest-auth = "^7.0.1"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 